Moved constant to generic module.
ActionSetup is now back to being for actions at the top of the input file only as requested by @Giovanni
